requirements.txt
streamlit
pandas
firebase_admin
flask
requests
python-dotenv